refactor: split the per-site catalog apis types into domain files (#1615)
