+++ site-config/brc-analytics/local/entity/priorityPathogen/priorityPathogenTop.ts
@@ -1,5 +1,5 @@
 import * as V from "@/viewModelBuilders/catalog/brc-analytics-catalog/common/viewModelBuilders";
-import { type Outbreak } from "@brc/apis/types";
+import { type Outbreak } from "@brc/apis/outbreak";
 import { BackPageHero } from "@databiosphere/findable-ui/lib/components/Layout/components/BackPage/components/BackPageHero/backPageHero";
 import {
   type ComponentConfig,
