+++ site-config/brc-analytics/local/index/common/column/columnRegistry.ts
@@ -1,7 +1,5 @@
-import {
-  type BRCDataCatalogGenome,
-  type BRCDataCatalogOrganism,
-} from "@brc/apis/types";
+import { type BRCDataCatalogGenome } from "@brc/apis/assembly";
+import { type BRCDataCatalogOrganism } from "@brc/apis/organism";
 import { type ColumnConfig } from "@databiosphere/findable-ui/lib/config/entities";
 import * as COLUMN_DEFS from "./columnDefs";
