+++ site-config/ga2/local/entity/organism/organismMainColumn.ts
@@ -4,7 +4,7 @@ import {
   type ComponentConfig,
   type ComponentsConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
-import { type GA2OrganismEntity } from "@ga2/apis/types";
+import { type GA2OrganismEntity } from "@ga2/apis/organism";
 import { Main as OrganismViewMain } from "@ga2/views/OrganismView/components/Main/main";
 
 export const organismMainColumn: ComponentsConfig = [
