+++ site-config/ga2/local/index/genome/columnDefs.ts
@@ -32,7 +32,7 @@ import {
   type ColumnConfig,
   type ComponentConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
-import { type GA2AssemblyEntity } from "@ga2/apis/types";
+import { type GA2AssemblyEntity } from "@ga2/apis/assembly";
 import { OrganismAvatar } from "@ga2/components/OrganismAvatar/organismAvatar";
 import { AnalyzeGenome } from "@repo/shared/components/Table/components/TableCell/components/AnalyzeGenome/analyzeGenome";
 import { LevelCell } from "@repo/shared/components/Table/components/TableCell/components/LevelCell/levelCell";
