+++ site-config/ga2/local/index/genome/columns.ts
@@ -1,5 +1,5 @@
 import { type ColumnConfig } from "@databiosphere/findable-ui/lib/config/entities";
-import { type GA2AssemblyEntity } from "@ga2/apis/types";
+import { type GA2AssemblyEntity } from "@ga2/apis/assembly";
 import {
   ACCESSION,
   ANALYZE_GENOME,
