+++ sites/brc-analytics/views/EntityView/assembly/components/Side/brc/types.ts
@@ -1,4 +1,4 @@
-import { type BRCDataCatalogGenome } from "@brc/apis/types";
+import { type BRCDataCatalogGenome } from "@brc/apis/assembly";
 
 export interface Props {
   assembly: BRCDataCatalogGenome;
