+++ sites/brc-analytics/views/PriorityPathogenView/components/ResourcesSection/types.ts
@@ -1,4 +1,4 @@
-import type { OutbreakResource } from "@brc/apis/types";
+import type { OutbreakResource } from "@brc/apis/outbreak";
 import type { SectionProps } from "@brc/views/PriorityPathogenView/ui/Section/types";
 
 export interface Props extends Omit<SectionProps, "children"> {
