+++ sites/ga2/apis/organism.ts
@@ -1,0 +1,21 @@
+import { type GA2AssemblyEntity, type ImageData } from "./assembly";
+
+export interface GA2OrganismEntity {
+  assemblyCount: number;
+  assemblyTaxonomyIds: string[];
+  genomes: GA2AssemblyEntity[];
+  image: ImageData;
+  maxScaffoldN50: number | null;
+  ncbiTaxonomyId: string;
+  taxonomicGroup: string[];
+  taxonomicLevelClass: string;
+  taxonomicLevelDomain: string;
+  taxonomicLevelFamily: string;
+  taxonomicLevelGenus: string;
+  taxonomicLevelKingdom: string;
+  taxonomicLevelOrder: string;
+  taxonomicLevelPhylum: string;
+  taxonomicLevelSpecies: string;
+  thumbnailUrl: string;
+  tolId: string;
+}
