+++ sites/ga2/views/EntityView/assembly/components/Side/ga2/components/AssemblyDetails/types.ts
@@ -1,4 +1,4 @@
-import { type GA2AssemblyEntity } from "@ga2/apis/types";
+import { type GA2AssemblyEntity } from "@ga2/apis/assembly";
 
 export interface Props {
   assembly: GA2AssemblyEntity;
