+++ tests/analysisMethodCatalogs.utils.test.ts
@@ -7,7 +7,7 @@ import type { WorkflowCategory } from "@repo/shared/apis/workflow";
 import { buildAssemblyWorkflows } from "@repo/shared/views/AnalyzeWorkflowsView/components/Main/utils";
 import { DIFFERENTIAL_EXPRESSION_ANALYSIS } from "@repo/shared/workflow/differentialExpressionAnalysis";
 import { WorkflowCategoryId } from "../catalog/schema/generated/schema";
-import type { BRCDataCatalogGenome } from "../sites/brc-analytics/apis/types";
+import type { BRCDataCatalogGenome } from "../sites/brc-analytics/apis/assembly";
 
 describe("buildAssemblyWorkflows", () => {
   const WORKFLOW_CATEGORIES: WorkflowCategory[] = [
